fix changes

+++ database/migrations/2024_10_21_111350_drop_user_id_field_from_collections_table.php
@@ -6,7 +6,8 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class () extends Migration {
+return new class extends Migration
+{
     /**
      * Run the migrations.
      */
@@ -34,8 +35,10 @@ return new class () extends Migration {
             }
 
             Schema::table('collections', function (Blueprint $table) {
-                $table->dropIndex(['user_id']);
-                $table->dropColumn(['user_id']);
+                if (DB::getDriverName() !== 'sqlite') {
+                    $table->dropIndex(['user_id']);
+                    $table->dropColumn(['user_id']);
+                }
             });
         }
